Add a grace period before hiding the quality warning
If the quality was seen as very bad for a split second this showed and
quickly hidden again the quality warning, which is annoying as it may
not be possible to even read the tooltip. To mitigate that now the
quality warning is shown at least for 3 seconds before hiding it again.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>