Dimm own avatar and video when the sender connection quality is bad
The dimming amount is the same used when a participant is disconnected,
although in that case there will be also a loading spinner which is not
shown here.

The dimming is meant to replace the warning icon as a way to make the
user aware of an issue with the connection when the tooltip is hidden,
so an ARIA label is provided too.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>